requirements.txt

gradio
python-docx
python-magic
sentence-transformers
faiss-cpu
regex
